standalone/corvus-data-generator/requirements.txt
pandas==2.2.1
PyYAML==6.0.1
requests==2.31.0
pydantic==2.6.4